requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
numpy==1.26.3
scipy==1.12.0
scikit-learn==1.6.1
torch==2.5.1+cu124
transformers==4.45.2
huggingface_hub==0.25.0
adapters==1.0.1
click==8.1.7
ruamel.yaml==0.18.6
jinja2==3.1.3
tqdm==4.67.1
logzero==1.7.0
